Product Combination : * Merge files * delete already merged file
